Split specs.py into Mixer and Midi
