+++ ludwig/boards/gld.py
@@ -1,5 +1,4 @@
-from ludwig import mixer
-from ludwig.specs import Midi, Mixer
+from ludwig import mixer, Mixer, Midi
 from rtmidi.midiconstants import NOTE_ON
 from ludwig.types import uint1, uint2, uint3, uint4, uint7, uint8
 from pydantic import conint
